requirements.txt
blinker==1.6.3
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
envsubst==0.1.5
Flask==3.0.0
Flask-Cors==4.0.0
flask-swagger==0.2.14
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pymongo==4.5.0
PyYAML==6.0.1
watchdog==3.0.0
Werkzeug==3.0.0
waitress==2.1.2
pytimeparse==1.1.8
schedule==1.2.1